feat(dal,sdf,web,cyc,ver) remove resource entity
