+++ lib/cyclone-client/src/client.rs
@@ -756,20 +756,17 @@ mod tests {
                     properties: serde_json::json!({"salt": "n", "peppa": "pig"}),
                     system: None,
                     kind: ComponentKind::Standard,
-                    resource: None,
                 },
                 parents: vec![
                     ComponentView {
                         properties: serde_json::json!({}),
                         system: None,
                         kind: ComponentKind::Standard,
-                        resource: None,
                     },
                     ComponentView {
                         properties: serde_json::json!({}),
                         system: None,
                         kind: ComponentKind::Standard,
-                        resource: None,
                     },
                 ],
             },
@@ -850,20 +847,17 @@ mod tests {
                     properties: serde_json::json!({"salt": "n", "peppa": "pig"}),
                     system: None,
                     kind: ComponentKind::Standard,
-                    resource: None,
                 },
                 parents: vec![
                     ComponentView {
                         properties: serde_json::json!({}),
                         system: None,
                         kind: ComponentKind::Standard,
-                        resource: None,
                     },
                     ComponentView {
                         properties: serde_json::json!({}),
                         system: None,
                         kind: ComponentKind::Standard,
-                        resource: None,
                     },
                 ],
             },
@@ -943,7 +937,6 @@ mod tests {
                     properties: serde_json::json!({ "si": { "name": "Aipim Frito" } }),
                     system: None,
                     kind: ComponentKind::Standard,
-                    resource: None,
                 },
                 codes: Vec::new(),
                 parents: Vec::new(),
@@ -1035,7 +1028,6 @@ mod tests {
                     properties: serde_json::json!({ "si": { "name": "Aipim Frito" } }),
                     system: None,
                     kind: ComponentKind::Standard,
-                    resource: None,
                 },
                 codes: Vec::new(),
                 parents: Vec::new(),
@@ -1125,7 +1117,6 @@ mod tests {
                 properties: serde_json::json!({ "si": { "name": "Aipim Frito" } }),
                 system: None,
                 kind: ComponentKind::Standard,
-                resource: None,
             },
             code_base64: base64::encode(
                 r#"function checkit(component) {
@@ -1213,7 +1204,6 @@ mod tests {
                 properties: serde_json::json!({ "si": { "name": "Aipim Frito" } }),
                 system: None,
                 kind: ComponentKind::Standard,
-                resource: None,
             },
             code_base64: base64::encode(
                 r#"function checkit(component) {
@@ -1297,7 +1287,6 @@ mod tests {
             properties: serde_json::json!({ "si": { "name": "Ablublubé"}}),
             system: None,
             kind: ComponentKind::Standard,
-            resource: None,
         };
         let req = CodeGenerationRequest {
             execution_id: "1234".to_string(),
@@ -1327,7 +1316,7 @@ mod tests {
                 assert_eq!(
                     serde_json::from_str::<serde_json::Value>(&output.message)
                         .expect("Unable to serialize output to json"),
-                    serde_json::json!({"properties": { "si": { "name": "Ablublubé" } }, "system": null, "kind": "standard", "resource": null })
+                    serde_json::json!({"properties": { "si": { "name": "Ablublubé" } }, "system": null, "kind": "standard" })
                 )
             }
             Some(Ok(unexpected)) => panic!("unexpected msg kind: {:?}", unexpected),
@@ -1387,7 +1376,6 @@ mod tests {
             properties: serde_json::json!({ "si": { "name": "Ablublubé"}}),
             system: None,
             kind: ComponentKind::Standard,
-            resource: None,
         };
         let req = CodeGenerationRequest {
             execution_id: "1234".to_string(),
@@ -1417,7 +1405,7 @@ mod tests {
                 assert_eq!(
                     serde_json::from_str::<serde_json::Value>(&output.message)
                         .expect("Unable to serialize output to json"),
-                    serde_json::json!({"properties": { "si": { "name": "Ablublubé" } }, "system": null, "kind": "standard", "resource": null })
+                    serde_json::json!({"properties": { "si": { "name": "Ablublubé" } }, "system": null, "kind": "standard" })
                 )
             }
             Some(Ok(unexpected)) => panic!("unexpected msg kind: {:?}", unexpected),
