+++ lib/cyclone-core/src/component_view.rs
@@ -20,18 +20,12 @@ pub struct SystemView {
     pub name: String,
 }
 
-#[derive(Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
-pub struct ResourceView {
-    pub data: Value,
-}
-
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ComponentView {
     pub system: Option<SystemView>,
     pub kind: ComponentKind,
     pub properties: Value,
-    pub resource: Option<ResourceView>,
 }
 
 impl Default for ComponentView {
@@ -40,7 +34,6 @@ impl Default for ComponentView {
             system: Default::default(),
             kind: Default::default(),
             properties: serde_json::json!({}),
-            resource: None,
         }
     }
 }
