+++ lib/cyclone-core/src/lib.rs
@@ -30,7 +30,7 @@ mod workflow_resolve;
 pub use canonical_command::{CanonicalCommand, CanonicalCommandError};
 pub use code_generation::{CodeGenerated, CodeGenerationRequest, CodeGenerationResultSuccess};
 pub use command_run::{CommandRunRequest, CommandRunResultSuccess};
-pub use component_view::{ComponentKind, ComponentView, ResourceView, SystemView};
+pub use component_view::{ComponentKind, ComponentView, SystemView};
 pub use confirmation::{ConfirmationRequest, ConfirmationResultSuccess};
 pub use encryption_key::{EncryptionKey, EncryptionKeyError};
 pub use liveness::{LivenessStatus, LivenessStatusParseError};
