+++ lib/dal/src/builtins/func/resourceExistsConfirmation.js
@@ -1,6 +1,9 @@
 async function exists(component) {
-  if (component.resource === undefined || component.resource === null
-      || component.resource.data === undefined || component.resource.data === null) {
+  if (component.properties.resource === undefined
+      || component.properties.resource === null
+      || component.properties.resource === ""
+      || component.properties.resource === "undefined"
+      || component.properties.resource === "null") {
     return {
       success: true,
       recommendedActions: ["create"]
