+++ lib/dal/tests/integration_test/mod.rs
@@ -30,7 +30,6 @@ mod qualification_check;
 mod qualification_prototype;
 mod qualification_resolver;
 mod read_tenancy;
-mod resource;
 mod schema;
 mod secret;
 mod socket;
